requirements.txt
langchain==1.2.10
langchain-openai==1.1.10
langsmith>=0.3.16
numpy>=1.26.0
openai==2.21.0
Pillow>=10.0.0
PyYAML>=6.0